Added support for Debian
The include task for installing OS specific packages has been changed to import the file based on the value of the OS family, so 'install_os_packages.yml' was renamed to 'install_RedHat_packages.yml'. Additionally, the variable files for each OS includes the variable 'pypi_server_bin' to represent the path for the 'pypi-server' executable. This was needed for the systemd service template, since pypi-server is installed to different locations on the different OSs. An additional test for the Debian platform was added, and documentation was updated to include the reference to the debian/jessie64 test.